refactored a bit
